Fix a bug where the job never runs unless manually triggered.
This bug is a result of misunderstanding the idle mode check.
Due to the quirks of the JobScheduler, we need to check the
preconditions every time the job runs. Thus, the job checked
if the system was idle.

The catch is that the system is not technically idle when
running idle jobs, as per PowerManager.isDeviceIdleMode().
Because we are scheduled during that window, the job never
actually runs.

Bug: 30034478
Change-Id: Ic6fa9922e372256a55a977a371e6e853ef877307