Fix a bug where the apps deletion type no longer populated.
This bug resulted from a recent change to how the ApplicationsState
works. It used to return a list of apps synchronously, but now only
works asynchronously after ag/1225943. The item title was populated
once we got the list of apps, but since we no longer received it,
the item title was no longer populated.

This patch does two things:
1. Use ApplicationsState asynchronously to properly populate.
2. Default to 0 clearable apps w/ 0 freeable bytes on initialization.

Bug: 30220563
Change-Id: I92ef1ca40b74b3d62673b828d6b8143680dd760a