Fix a bug where apps which should not be eligible for deletion were.
This rseulted from a logic error involving selecting the maximum of
the days since first install and days since last use. We want to
use what comes first: the install time or the last use. This should
be done by selecting the smallest amount of days of the two.

Instead, the largest of the two is chosen, causing apps which were
installed a while ago to trigger as eligible for deletion when
they should not be.

Bug: 30413088
Change-Id: I2178c0e995a8e3c4888593d43b46888f07916f79