AppDeletionPreference used app name instead of package name.
Using the label name can cause conflicts if multiple apps have the
same label. Package names must be unique.

Prior to this change, AppDeletionPreference used the app label as its
key. Because the key was not unique, this caused a bad interactions with
the other Deletion Helper parts. It causes duplicated entries in the
Deletion Helper and invalid behavior when selecting which packages to
delete.

Bug: 31089224
Test: m -j32 RunStorageManagerRoboTests
Change-Id: I6dd9627818d5a154d04d17ae2324bb65865148a7