Harden the PreferenceListCache against invalid keys.
If multiple preferences have the same key, we can no longer ensure
the PreferenceListCache can return the correct preference for the
duplicated key. If the preference has no key, we similarly cannot
guarantee correctness.

By throwing an exception in this case, we stop an app from getting
into an invalid state.

Bug: 31158678
Test: m -j32 RunStorageManagerRoboTests
Change-Id: I371b316e72328450b569293ec8735734a85b112a