Fix a bug where Deletion Helper could crash on deletion size callback.
This resulted from cases where the user very quickly navigates away
from the Deletion Helper while it is still loading. When the clearable
storage information completes, the callback executes and attempts to
update a label on the Activity. If the Activity no longer exists,
then the callback throws a NPE and crashes.

Bug: 31602093
Test: Manually reattempted the repro case.
Change-Id: I505f8a7558b58c94fdb88a558c6e9c0814e45210