Fix a bug where the ASM upsell never stops showing.
This occurred because when the delay says to "NEVER" show again,
which is coded as -1, is added to the current time, rather than
overriding it. This meant that the "NEVER" case turned into,
effectively, an always case.

Bug: 31979661
Test: m RunStorageManagerRoboTests
Change-Id: Ib503f83c90ec3423e01c242ab105a800f6794bc5
(cherry picked from commit 9849f1a2caeadd773da29b5a9fed9e486e4c0be4)