Fix a bug where the new 1 year usage stats query was broken.
This bug was not caught by the robotests because they were completely
mocked out and did not care about the values being passed into them.
This bug was missed in manual testing due to me not paying enough
attention to the apps list that was showing up in Deletion Helper and
not seeing it was more permissive than it should have been.

Bug: 34933090
Test: New Robotest
Change-Id: I28666bd6ead4cbfb4145ae2e7f3847a1d6d85638