Don't make 0B freeable.
The apps deletion provider was incorrectly saying it had clearable
bytes when it did not in its callback. As a result, toggling it on
and off to force a callback would make the view believe it had bytes
to clear.

The Free button itself, though, got the correct 0B number using a
different calculation.

This patches the invalid clearable bytes calculation for app deletion.

Change-Id: Ia3033806e4fc12b7a51d181fcc5e935cf0f7f580
Fixes: 35946083
Test: StorageManager Robotest