Add canary warnings for when UsageStats may be wrong.
This will assist with getting bug reports about when the
Deletion Helper is being overzealous. This patch adds an alternate
method of determining the usage stats and compares it to the original
method. If they mismatch, it warns.

Furthremore, we also report information about all apps which are
reported to be clearable.

Bug: 36906669
Test: Manual -- this patch should only cause logging.
Change-Id: I95d3fcade5bf6bf1ac5f2c777ae0ba8a5c31fbfe