Use Intermediary state in deletion types
Having only a loaded content/no content state
was making so that the DeletionHelperActivity
would occasionally change states when it
shouldn't. This CL changes the states to
'unfinished', 'done with no content', and
'done with content.' This will allow the
activity to know for sure if the
DeletionTypes are done loading and found
nothing, or if they are still just searching
for content.

Test: Robotests
Bug: 37434948
Change-Id: If7cfb09e94e2710bc1293455caeafd3a627ee2c4