Fix NPE when listing recent files
When there are no recent items in storage, mDownloadsDeletion gets set
to null in setupEmptyState. Later on when we try to load the list of
recent files, we deference this null object, causing a crash.

Fixes: 67104773
Change-Id: I9e2bfff7ad13a4ecd72758d52bf4194a425f3b52
Test: make RunStorageManagerRoboTests

Verify that without the fix, NPE in the newly added test.
After fix, all tests pass