Fix test to use better flag.
Previously, the show_opt_in and enabled flags were conflated to
basically mean the same thing. Because we are now distinctly making
show_opt_in refer to showing a toggle to opt into the feature in SUW and
enabled to mean the default enabled status, this unit test needed to be
updated.

Additionally, the unit test has been renamed for clarity.

Bug: 122461924
Test: m RunStorageManagerRobotests
Change-Id: Ifc61efcc65af5becdded161048f5606c09d83013
Merged-In: Ifc61efcc65af5becdded161048f5606c09d83013